Migrate from deprecated `yaml` packages (#12895)
* Switch packages that can be migrated as-is

* Switch packages to `go.yaml.in/yaml/v4` with options

This is needed, as the v2 does not do stable marshalling in this repository.
When switching to v3/v4 of the new upstream, the defaults for indentation is 4 and arrays are additionally indented.
This is why we need to use this custom encoder here

* Adapt tests

This was a fun one...
The old package did some random newlines, whilst the new ones cannot do this.

Once again, we cannot use the new v2, because of the unstable sorting issue.

There is an _internal_ option in the parser to set a maximum width, but these are unexported options. ¯\_(ツ)_/¯

* Remove forked code

* `go mod tidy`

* Update skaffold deps

* Implement PR feedback

* Use `go.yaml.in/yaml/v4` instead of `go.yaml.in/yaml/v3`