Introduce `ipFamilies` in `Garden` API for runtime cluster
The kind cluster in the local dual-stack setup has two IP families, so
we also need to reflect this - mostly, in order to properly configure
the istio used for exposing the virtual garden -> see next commit.

Note that this "ipFamily"/dual-stack completion is not complete for the
`Garden` with this enhancement. It is just one puzzle piece to make the
local setup work. If we really want to support dual-stack runtime
clusters, we need to put some follow-up effort into this story.