Add test to demonstrate Prometheus instances are checked sequentially
Querying Prometheus for health check alerts can become expensive. In the
worst case, if a Prometheus instance is down, the regular condition
check will block waiting for it to reply until the connection times out.
The more Prometheus pods are down or have issues, the longer it will
take for the condition check to complete.

This commit adds a unit tests to validate the time it takes to query 6
Prometheus pods in total: 2 managed Prometheus with 3 replicas each.
Such test will be used in the following commits to show how this
execution time reduces as Prometheus instances are checked in parallel.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>