Fix alertmanager peer discovery (#12991)
An upstream change introduced named ports for the service created by the
prometheus operator. In turn, that caused the NetworkPolicy controller to
generate network policies with a different name and pod selector, while
the alertmanager instances still had the old label. This commit adapts
the alertmanager custom resource to include the new name.

Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>