Reconcile garden resource as part of operator-seed-up
This change is necessary for the garden operator to pick up the recently
installed local seed and create the corresponding configurations in the
garden Prometheus to federate from the seed's aggregate Prometheus.
Without this, the federation between these two Prometheus instances
can't be validated in the local setup.

This is a side effect of bootstrapping and there might be other aspects
that will benefit from reconciliation.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>