--- Introduce new health check rules in Prometheus ---
This empty commit marks the beginning of a new section in the commit
history, where upcoming commits will introduce recording rules for
Prometheus to report the current health of the monitoring stack.

It introduces two types of recording rules:

1. "healthcheck" with a distinguishing "task" label.
2. "healthcheck:alert" is a count of "healthcheck" to indicate whether
   there are any failing or flapping health checks.

The "task" label is used to distinguish between different health check
rules and avoid problems when these time series are aggregated. Namely,
Prometheus removes the name of a time series when using aggregation
functions. The "task" label ensures that the label set is sufficient to
distinguish different time series.

The value of the "healthcheck" time series is not significant.
Analogical to the Prometheus ALERTS builtin metric, values only indicate
a failing health check of the given type, while the absence of a value
indicates health.

The "healthcheck:alert" time series uses a "count" aggregation to emit
either 0 or 1 time series. To avoid redundancy, the task label is
removed by the "count" aggregation by design.

The "healthcheck:alert{flapping="true"}" time series specifically
indicates a flapping "healthcheck:alert". By separating this dedicated
time series from regular "healthcheck:alert", it helps to investigate
potential issues and prevents frequent updates to the shoot, seed, or
garden resource status caused by flapping health checks, which would
otherwise result in frequent writes to etcd.

The care controller can query for "healthcheck:alert" to determine
whether there are any failing health checks, flapping or not, and then
reflect a corresponding condition in the status of the shoot, seed or
garden resource.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>