--- Introduce new health check rules in Prometheus ---
This empty commit marks the beginning of a new section in the commit
history, where upcoming commits will introduce recording rules for
Prometheus to report the current health of the monitoring stack.

Each rule is designed to return 1 when a failure or anomaly is detected,
and nothing when healthy, ensuring consistent semantics across all
checks.

For example, the rule `up == 0 == bool 0` first filters for targets that
are down (`up == 0`), and then uses `== bool 0` to transform the result
into a value of 1.

Each health check is represented by a meta health rule labeled with
`type="health:meta"`. These meta health rules are combined by the
`health:failing` rule to provide a single, seamless view of health
checks. For instance, if a scrape or federation job is down, multiple
meta health rules may report an issue: first, because the job is down,
and second, because there are no samples flowing in. Thanks to
`health:failing`, these are combined so that only the job down part
triggers, while inhibiting the rest. The `health:failing` rule is also
especially useful for promtool unit tests, as it verifies that all meta
health rules are considered.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>