Use a variable of type struct for Prometheus service information
This is a small refactor that will come in handy in the following
commits, where the service configuration of the longterm Prometheus
needs to be reworked to change the web and cortex service ports. This
allows for grouping of related information. For now, only the web port
is moved, but in the next commit the cortex port is also added.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>