Rework service configuration for longterm Prometheus
Currently, the longterm Prometheus is exposed on 80->9091. This commit
changes this and extends the service ports to expose both the Prometheus
and Cortex ports:

80->9090
81->9091

With this change, the network controller will create network policies to
access the longterm Prometheus on port 9090, exactly as it currently
does for other Prometheus.

The longterm Prometheus ingress has been changed to keep mapping to the
Cortex frontend, now on the service port 81 as mentioned above.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>