[GEP-28] Implement `ShootAuthorizer` and `ShootRestriction` to control permissions of shoot `gardenlet`s (#12978)
* Prefactor: Move `seedidentity` pkg to `gardenletidentity/seed`

In the next commit, we introduce a similar package for gardenlet shoot
identities.

* Shoot Identity package

* Prefactor: Move auth helper functions out to make them reusable

The shoot authorizer needs some of them as well.
Next commits makes the extract helper functions more generic.

* Prefactor: Make auth helper functions generic

Previously, they were hard-coded to the "seed" scenario, but soon we
will need to use them for the "shoot" scenario

* Create new shoot authorizer handler in `gardener-admission-controller`

* Prepare `graph` package for handling for autonomous shoots

(There is nothing now/in the scope of this PR, but it will naturally be
extended as development progresses.)

* Register shoot authorizer webhook in `virtual-garden-kube-apiserver`

* Create new shoot restriction handler in `gardener-admission-controller`

* Address PR review feedback

* Address PR review feedback

* Address PR review feedback