[GEP-28] `gardenadm bootstrap`: Deploy bootstrap `DNSRecord` (#12981)
* Explicitly specify `Shoot.spec.dns` in gardenadm manifests

`Shoot.spec.dns.domain` will be used as the shoot's external domain (there is no internal domain for autonomous shoots).

* Add helper for distinguishing gardenadm scenarios

* Add `DNSRecord` to calculation of required extensions

* Clarify naming of `gardenadm discover` flags

* Initialize botanist with DNS info

With this commit, the botanist no longer uses a hard-coded internal domain. Instead, the external domain is read from `Shoot.spec.dns.domain`.
There is no internal domain for autonomous shoots, only an external domain.
However, it might be managed outside gardener (high-touch). In this case, the field is only for informing `gardenadm` about the domain for secret generation, etc.

* Make `Shoot.InternalClusterDomain` optional

Similar to `ExternalClusterDomain`, `InternalClusterDomain` is fully optional now.
Hence, we make it a pointer as well and handle a `nil` pointer in the botanist code.

* Prefactor: store list of machines for reuse

This list will be reused for creating the `DNSRecord`.
Also, treat unknown address types as lowest priority instead of highest (same as `InternalIP`).

* Deploy bootstrap `DNSRecord`

* Adapt provider-local `DNSRecord` controller

This commit
- simplifies the `DNSRecord` actuator code and unit tests
- configures CoreDNS to return `DNSRecord.spec.values` directly for autonomous shoots instead of rewriting to the istio-ingressgateway name
- ensures `DNSRecords` of autonomous shoots are not deleted on migration

* Drop `external` from domain of autonomous shoot clusters

* Address review feedback

* Fix logic to skip `DNSRecord` deletion

In the previous `DNSRecord` controller commit, I unintentionally removed the skip logic from the `DNSRecord` deletion.
This lead to errors when deleting a `Garden`:
```yaml
    lastOperation:
      description: "1 error occurred:\n\t* task \"Destroying DNSRecords for virtual
        garden cluster and ingress controller\" failed: 3 errors occurred:\n\t* Failed
        to delete DNSRecord garden/wildcard-ingress-runtime-garden-local-gardener-cloud-local:
        Error deleting DNSRecord: namespaces \"gardener-extension-provider-local-coredns\"
        not found\n\t* Failed to delete DNSRecord garden/api-virtual-garden-local-gardener-cloud-local:
        Error deleting DNSRecord: namespaces \"gardener-extension-provider-local-coredns\"
        not found\n\t* Failed to delete DNSRecord garden/gardener-virtual-garden-local-gardener-cloud-local:
        Error deleting DNSRecord: namespaces \"gardener-extension-provider-local-coredns\"
        not found\n\n\n\n Operation will be retried."
```

* Unexport `ControlPlaneMachines`

* Address review feedback