Introduce `RequeueReconciler` if and `RateLimitedRequeReconcilerAdapter`
This commit adds a `RequeueReconciler` interface that can be used to
create reconcile logic that can explicitly state whether a requeue
should be performed.

This is added to support the same behaviour as the now deprecated
`reconcile.Result{Requeue: true}` from controller-runtime - when
`reconcile.Result{Requeue: true, RequeueAfter: 0}` was returned
from a reconciliation without any `RequeueAfter` duration, the
reconciliation would be requeued based on a rate limiter.

The `RequeueReconciler` interface can be used with the
`RateLimitedRequeueReconcilerAdapter` function which provides rate
limiting of requeues and adapts the interface so that it can be used
as a `reconcile.Reconciler` interface