Add timeouts to remote command execution
This adds context handling to `ssh.Connection.Run*`.
If the context is canceled (e.g., due to a timeout), we send `SIGINT` to the remote process.
If the process doesn't terminated gracefully within 30s, we send `SIGKILL`.
With this, we can rely on the remote process to be finished when `Run*` returns.