--- Expose longterm Prometheus on port 9090 ---
Previous commits introduced logic to query Prometheus for health issues
on port 9090. Such port is unfortunately not configurable by the
Prometheus operator.

In such setup, the longterm Prometheus becomes an edge-case because it
is only exposed on port 9091, where a Cortex sidecar container runs.
Services expose other Prometheus on 80->9090, but the longterm
Prometheus on 80->9091. In consequence, the Gardener network policy
controllers do not create network policies to access the longterm
Prometheus on port 9090.

While we could create a specific network policy to access 9090, the
following commits update the longterm Prometheus service to also expose
9090. This way, the longterm Prometheus service aligns with other
Prometheus on 80->9090, while adding a new port for Cortex 81->9091.

The corresponding longterm Prometheus ingress to port 80 is now changed
to use port 81.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>