Rework longterm Prometheus service to expose port 9090
Previous commits query Prometheus for health issues on port 9090, which
is not configurable in the Prometheus operator.

Longterm Prometheus is currently only exposed on port 9091 for the
Cortex sidecar. Other Prometheus expose 80->9090, but longterm exposes
80->9091. This means the network policy controller does not create
policies for port 9090 on longterm Prometheus.

Rather than adding a special network policy, the service is reworked to
also expose 9090. The service now has 80->9090 like other Prometheus,
plus a new port 81->9091 for Cortex. This way, the network controller
automatically creates the right policies.

The longterm Prometheus ingress is updated to point to the Cortex
frontend on the new service port 81.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>