[GEP-28] Enable bootstrap procedure and `Gardenlet` controller for self-upgrades (#13034)
* Prefactor: `RequestKubeconfigWithBootstrapClient` does not return seed name

The seed name is already an input parameter - there is no point in also
adding as output parameter.

* `gardenlet` bootstrap procedure creates CSR for autonomous shoot

Common name: `gardener.cloud:system:shoot:<namespace>:<name>`
Groups: `gardener.cloud:system:shoots`

It reads the `shoot-info` `ConfigMap` in the `kube-system` namespace in
order to get information about the shoot it is running in.

For now, we don't delete the bootstrap auth secret because we don't have
permission for it yet. This will be added in a subsequent commit.

* `gardener-controller-manager` auto-approves `gardenlet`'s CSRs for autonomous shoots

* `gardenlet` starts garden cluster and registers for self-upgrades

* Prefactor: Allow providing values for label/field selectors

Needed to restrict `gardenlet`'s read requests for `Gardenlet` resource
to the resource they are responsible for (the `Gardenlet` resources are
prefixed with `autonomous-shoot-<obj-name>`). Previously the authorizer
was always checking for `<obj-name>` only.

* Handle CSRs and `Gardenlet`s in shoot authorizer

* Handle CSRs and `Gardenlet`s in shoot restriction

* Field selector for `Gardenlet` watches

* `gardenlet` first creates `Gardenlet` resource

- Otherwise, the authorizer might reject it when there is no Seed/Shoot
  yet

* Adapt certificate manager to prevent confusing logs

- now it knows about autonomous shoots
- and it can report proper events there

* `Gardenlet`s for autonomous shoots are auto-updated

via Skaffold

`gardener-operator` is never responsible for deploying `gardenlet`s into
autonomous shoots (hence, the `OperatorResponsible` predicate of its
`Gardenlet` controller now always returns `false` for autonomous
shoots). This should only happen via `gardenadm connect`.

* Enable `Gardenlet` controller for self-upgrade

* Only set seed name when `SeedConfig` is set

Previously, the `gardenlet` running in an autonomous shoot would try to
set the seed name if it starts up and reconciles its `Gardenlet`
resource.
In the previous implementation, the `obj` type was checked - but in
above case, it is a `seedmanagementv1alpha1.Gardenlet`.

* Restrict namespace for `Gardenlet` resources

Otherwise, a shoot gardenlet would have access to all `Gardenlet`
resources with name `autonomous-shoot-<gardenlet-name>`, even if they
are in other namespaces:

```
$ kubectl -n default     get gardenlet --field-selector metadata.name=autonomous-shoot-root
No resources found in default namespace.
$ kubectl -n kube-system get gardenlet --field-selector metadata.name=autonomous-shoot-root
No resources found in kube-system namespace.
```

* Enable bootstrap token secret deletion after bootstrap

- gardenlet deletes its bootstrap token secret after it aquired
  a client certificate via CSR
- the shoot authorizer checks if the to-be-deleted bootstrap secret
  belongs to this gardenlet

* Adapt e2e test

* Address PR review feedback

* Address PR review feedback