(aggregate) Ignore shoot-prometheus job for health checking
Prometheus health check in a seed is reflected by the
SeedSystemComponentsHealthy condition. The gardenlet blocks seed
operations such as shoot reconciliation when this condition is false.
This is acceptable when there are health issues with the cache,
aggregate or seed Prometheus as this could indicate a problem with the
seed itself.

However, aggregate Prometheus has a job to federate from shoot
Prometheus. Issues with the shoot control plane itself could take such
job down. In such circumstances, it is not acceptable to block seed
operations for other shoots.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>