Add opt-out functionality for zone based istio ingress gateways (#12956)
* feat: introduce zonalIngress field to seed settings

Signed-off-by: cerealsnow <13483299+cerealsnow@users.noreply.github.com>

* feat: restrict istio ingress gateway creation based on zonalIngress field

Signed-off-by: cerealsnow <13483299+cerealsnow@users.noreply.github.com>

* feat: cleanup namespaces for zonal istio ingress gateways

* feat: select global ingress gateway if zonal gateway is disabled

Shoots with failureTolerance != zone use the zonal istio ingress gateways. For seeds with disabled zonal gateways, we need to select the global gateway instead.

* feat: use global gateways for local dns records if zonal gateways are disabled

* docs: document usage of zonalIngress field in Seed API

* fix: align pull-gardener-unit tests with managed istio-ingress namespace

The istio-ingress namespace is now provisioned as a managed resource, which caused test failures. Added validation to ensure the namespace is deployed as a managed resource.

* fix: remove cleanup of namespaces for zonal istio ingress gateways

The deployment and cleanup of namespaces for zonal istio ingress
gateways is now handeled by the GRM, so a manual cleanup is not
necessary anymore.

* fix: align pull-gardener-integration tests with zonalIngress Seed API field

* docs: add doc string for zonalIngress field in Seed API to external types

* feat: add output of make generate

* feat: safeguard deletion of unused zonal istio ingress gateways

When an operator disables zonal istio ingress gateways, their cleanup is
now postponed
until it is verified that no Shoots are still referencing them. This
prevents the control planes of affected Shoots from becoming unreachable
until the next reconciliation.

Signed-off-by: cerealsnow <13483299+cerealsnow@users.noreply.github.com>

* feat: return error from AreZonalGatewaysInUse and handle it properly

Signed-off-by: cerealsnow <13483299+cerealsnow@users.noreply.github.com>

* fix: handle NoMatchError for Gateway CRD

Signed-off-by: cerealsnow <13483299+cerealsnow@users.noreply.github.com>

---------

Signed-off-by: cerealsnow <13483299+cerealsnow@users.noreply.github.com>