Removed seedmanagement/applyconfiguration and settings/applyconfiguration.
In the past, the corresponding packages were autogenerated, but since
PR#8851 they are no longer updated. It does not look like we actually
need them. Hence, removing them seems like the better option compared to
re-enabling autogeneration.
As a plus, this enables us to get rid of a direct dependency.