Ensure `ShootMutator` is executed before the `ShootValidator` admission plugin
The current order of execution is such that `ShootValidator` gets executed before `ShootMutator`.
If there is a field defaulted in `ShootMutator` and later on validated for nil in the `ShootValidator`, then the `ShootValidator` would reject the field as it gets executed first.
With this commit, we execute the `ShootMutator` before `ShootValidator` so that the fields can be mutated before validation rejects them for being missing.