(managedseed) Add health check end-to-end tests for Prometheus
This adds an end-to-end test that deploys a failing health check rule in
all Prometheus instances deployed in the seed and validates the
corresponding conditions are successfully set to false.

The test execute after deploying a managed seed that can be tore down
upon test completion. This helps keeping these test isolated. Running
them directly in the seed deployed in the KinD cluster would be too
intrusive.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>