[GEP-28] Clean up some TODOs (#13290)
* Deploy machine CRDs in shoots with managed infrastructure

Previously, we only deployed the machine CRDs in `gardenadm bootstrap`.
With this, we also deploy it in `gardenadm init` for shoots with managed infrastructure.

* Disable high-availability-config webhook in bootstrap cluster

Disable the webhook in the kind cluster during `gardenadm bootstrap`, as it is not needed.

This also disables the webhook from the virtual garden, where it isn't needed either.

* Adapt CRDs unit test