--- Add end-to-end tests for new Prometheus health checks ---
This empty commit is a header for a section where commits will add new
end-to-end tests to validate the status conditions are set to false for
unhealthy Prometheus. To do that, the test will create a new `vector(1)`
health check rule to simulate an issue.

Prometheus deployed in the seed are tested using a managed seed that can
be tore down upon test completion. This helps keeping these test
isolated. Running them directly in the seed deployed in the KinD cluster
would be too intrusive.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>