Update Istio dashboards to accurately reflect the total resource usage
Prior to this change, a bug with stacked graphs caused Plutono to show
only the resource usage of pods that were running at the end of the query
time range. For time ranges longer than the default 30m, this caused the
visual effect of istio ingressgateway resource consumption rising when pods
were replaced.

This commit removes the stacking and introduces a `_total` metric that
sums over all *selected* istio ingressgateway pods. It can still be hidden
by shift-clicking it in the legend below the graph.