Refactor CheckManagedPrometheuses() into CheckPrometheuses()
This commit reworks the current approach used for Prometheus health
checking. Previously, Managed Resources were used as entry point for the
health check. Managed Resources were listed and searched for Prometheus
resources. This approach was suboptimal because Prometheus resources can
directly be listed, but it was done so in order to provide the ability
to use the "ignore" annotation to temporarily disable the health checks.

However, preserving the behavior of the "ignore" annotation turned out
to be irrelevant because it is only meant for dev scenarios. Keeping the
new Prometheus health check simpler is preferable. In fact, temporarily
disabling this feature can also be achieved by annotating the Prometheus
Managed Resource with "ignore" and deleting the problematic Prometheus
health recording rules.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>