Skip Prometheus health checks on IPv6 tests
Currently, the local setup used in the tests does not fully support
IPv6. We found there are cross-node network issues for shoots which are
not easy to solve. For now, we opt to skip Prometheus health checks on
IPv6 tests and tolerate unhealthy observability conditions when the
reason is "PrometheusHealthCheckDown".

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Johannes Scheerer <johannes.scheerer@sap.com>