Only deploy `DNSRecord` secret on `Destroy` if `SecretName` is set
In multiple places, we instantiate the `DNSRecord` component with neither `SecretName` nor `SecretData`.
The changed logic in this PR causes `DNSRecord.Destroy()` to try deploying a secret without a name.
For example, this causes the `Garden` deletion to fail with:
```
lastOperation:
  description: "1 error occurred:\n\t* task \"Destroying DNSRecords for virtual
    garden cluster and ingress controller\" failed: 3 errors occurred:\n\t* Secret
    \"\" is invalid: metadata.name: Required value: name or generateName is required\n\t*
    Secret \"\" is invalid: metadata.name: Required value: name or generateName
    is required\n\t* Secret \"\" is invalid: metadata.name: Required value: name
    or generateName is required\n\n\n\n Operation will be retried."
```

For simplicity, the `DNSRecord` component should keep support for this way of destroying existing `DNSRecord` objects.
Hence, we skip deploying the secret, if `SecretName` is empty.
This is basically the same behavior regarding `Destroy` as before the previous commit in this file in this PR.