Requeue projects on Shoot deletion events
This commit adds functionality to trigger project reconciliations on
shoot deletions if there aren't any shoots left in the project and the
project has a deletion timestamp.
This is so that the project cleanup can continue after all Shoots have
been deleted.
With this requeuing projects for reconciliations during deletion, if they
have Shoots remaining, with `Requeue: true` is no longer necessary.