Requeue ManagedResources when cleanup is completed
Adds a predicate that allows ManagedResources to be requeued after their
class has changed and the cleanup of resources has been completed by the
resource manager that was previously responsible.
With this requeuing with `Requeue: true` done by the new resource
manager while it waits for resources to be cleaned up is no longer
necessary.