Prefactor: use the same values for OSC components
This is needed for calculating the same OSC hash for the bootstrap OSC (only containing the gardenadm download unit) and the real/usual OSC.
Without the same hash, mcm will try to roll the control plane pool as soon as the `Worker` has been restored within the self-hosted shoot.

Also, the `controlPlaneBootstrap` implementation now reuses the `Data` type for conveniently returning the same type and values in
`WorkerPoolNameToOperatingSystemConfigsMap`.