Ensure node-agent has a user corresponding to `Machine` object
We need to pass the machine name from the mcm in the bootstrap cluster (instructed by `gardenadm bootstrap`)
to the `gardener-node-agent` (instructed by `gardenadm init`).
Otherwise, `gardener-node-agent` will use the host name instead of the machine name for requesting a client
certificate.
Later on, when activating the node-agent-authorizer, the client certificate user will not match the `Machine`
name causing `gardener-node-agent` to lose access and stop working.