feat(grm/webhook/vpainplaceupdates): add `mutated` label on handler completion
This commit adds a new "mutated" label and instruments the GRM webhook
logic to append it as a final stap of it's hander logic.

The main purpose of this label is to indicate that the resource has
already been processed by the webhook.

This new label will find purpose in a gardenlet initial migration logic
that looks into non-processed VPA resources and applies an empty update,
thus re-triggering the GRM webhook and migrating the pending VPA
resources.