Run registry and caches in docker containers instead of in kind pods
When bootstrapping self-hosted shoot clusters in docker containers (as a replacement for the kind cluster), we have a dependency on a local registry
where we push and pull gardener (gardenadm/provider-local) images to/from.
Hence, we cannot run the registry inside the cluster, but need to run it externally.

Running the local registry and registry caches via docker compose outside the kind cluster, has even more advantages:
- the registry image itself is now cached on the docker daemon of the host machine (previously it was pulled everytime on `make kind*-up`)
- it allows us to use dedicated hostnames instead of dedicated ports