Stop using internal IPs of nodes in service controller
Previously, we added the internal IPs of nodes to the `Service.status.loadBalancer.ips[]`.
With this, we exploited the kube-proxy shortcut for LoadBalancer IPs (it directly routes traffic to such IPs to the Services' clusterIP).
This was only necessary, because the load balancer IPs (`172.18.255.*`) overlapped with the kind docker network
and the traffic was not correctly routed to the load balancer IPs on the host machine's loopback interface.

With the previous commit, we correctly route traffic to the load balancer IPs (`172.18.255.*`) via the host machine.
Hence, we no longer need the shortcut added by kube-proxy.