Remove usages of `controller-runtime`'s deprecated `reconcile.Result{Requeue: true}` (#13052)
* Remove deprecated `Requeue` from extension reconciler

`Requeue` can be safely removed here as it would be true only if
`RequeueAfter` is also set.

* Remove deprecated `Requeue` from tokenrequestor reconciler

Removes usage of deprecated `Requeue` and the unnecessary `mustRequeue`
boolean as it was anyways only true when requeueAfter > 0

* Remove deprecated `Requeue` from workloadidentity reconciler

Removes usage of deprecated `Requeue` from the
`tokenrequestor/workloadidentity` reconciler and the unnecessary
`mustRequeue` boolean as it was anyways only true when requeueAfter > 0

* Remove deprecated `Requeue` from `ReconcileErr` function

* Remove unecessary deprecated `Requeue` from `ControllerInfos` test

* Replace deprecated `Requeue` from the garden reconciler

In this case requeue can be replaced with a very minimal duration.
Because it is expected to be only requeued once.

5 milliseconds was chosen because that was minimal requeue after
duration returned by the internal rate limiter previously, when
`Requeue: true` was used.

* Remove deprecated `Requeue` from garbagecollector reconciler

* Extract a constant for the default `RequeueAfter` time of 5ms

* Requeue projects on Shoot deletion events

This commit adds functionality to trigger project reconciliations on
shoot deletions if there aren't any shoots left in the project and the
project has a deletion timestamp.
This is so that the project cleanup can continue after all Shoots have
been deleted.
With this requeuing projects for reconciliations during deletion, if they
have Shoots remaining, with `Requeue: true` is no longer necessary.

* Requeue ManagedResources when cleanup is completed

Adds a predicate that allows ManagedResources to be requeued after their
class has changed and the cleanup of resources has been completed by the
resource manager that was previously responsible.
With this requeuing with `Requeue: true` done by the new resource
manager while it waits for resources to be cleaned up is no longer
necessary.

* Address review comments

* Refactor test with ordered `It`s