Update remote local setup with hosts for end-to-end tests and IPv6 support (#13436)
* Update /etc/hosts with new end-to-end test shoots

* Update /etc/hosts with entries for ci-e2e-kind-operator tests

* Update /etc/hosts with IPv6 entries

* Add command to run IPv6 on "prepare-operator-complete.sh" setup

Only the "prepare-operator-complete.sh" setup supports IPv6 currently.

* Deploy shoot-ipv6.yaml when IPv6 is used

* Support IPv6 for managed seed

This commit is analog to previous one in addition to introducing new
necessary files.

* Use new "credentialsBindingName"

The "secretBindingName" field is deprecated in favor of
"credentialsBindingName".

* Fix managed seed window panes

* Address review comments

This shifts IPv6 checks to occur during the creation of windows or
panes, and, thus, removing them from the window or pane command.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

---------

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>