Disable multi-value option for year variable
Selecting multiple values for year and month can lead to billing periods
without data. For instance, in January 2026, we might still have data
for December 2025. In such cases, two values become available for year
(2025 and 2026) and month (December and January). If both variables
are multi-value, billing periods like December 2026 and January 2025
become selectable. The former is in the future, while the latter
might be too old.

This commit addresses this issue by disabling the multi-value option for
the year variable and only providing available months for the selected
year. Comparing between two different years is not possible anymore in
one single dashboard, but this can be worked around by opening a
separate browser window.