Run registry and caches in docker containers instead of in kind pods (#13551)
* Run registry and caches in docker containers instead of in kind pods

When bootstrapping self-hosted shoot clusters in docker containers (as a replacement for the kind cluster), we have a dependency on a local registry
where we push and pull gardener (gardenadm/provider-local) images to/from.
Hence, we cannot run the registry inside the cluster, but need to run it externally.

Running the local registry and registry caches via docker compose outside the kind cluster, has even more advantages:
- the registry image itself is now cached on the docker daemon of the host machine (previously it was pulled everytime on `make kind*-up`)
- it allows us to use dedicated hostnames instead of dedicated ports

* Drop registry port mappings in kind cluster

* Adapt references to registry and caches

Co-Authored-By: Rafael Franzke <rafael.franzke@sap.com>

* Document `docker compose` requirement

* Fix e2e upgrade tests

* Don't stop registry on `make kind2-down`

* Clean up `TODOs` from preparing PRs

---------

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>