[GEP-28] `gardenadm connect`: Enable `Backup{Bucket,Entry}` controllers in `gardenlet` (#13519)
* `gardenadm connect` creates `Backup{Bucket,Entry}` in garden

- Previously, they were only computed/used in-memory during the
  `gardenadm init` run
- Now that we deploy a `gardenlet`, we want actual/"physical" resources
  in the garden cluster
- `gardenlet` has controllers for them to reconcile (to be enabled in
  the subsequent commits)

* Expose `.metadata.uid` and `.status.uid` in `shoot-info` `ConfigMap`

Since these values are part of the `BackupBucket` and `BackupEntry`
names for the self-hosted shoot, the `gardenlet` needs to know them in
order to be able to later restrict its cache to exactly these objects.

* Add `.status.uid` field selector for `Shoot`s

This will be helpful for the shoot restriction in
`gardener-admission-controller` to check if the creation of a generated
`BackupBucket` secret in the garden cluster belongs to the self-hosted
shoot that is attempting to do it.

* Prepare shoot authorizer/restriction for `BackupBucket` controller enablement

- allow creation of generated `BackupBucket` secret in garden cluster
- allow CRUD operations on `BackupBucket` resource
- enable resource dependency graph handling

* Handle `Secret`/`ConfigMap` creation via `gardenadm` in shoot restriction handler

Now that the `shoot-restriction` is properly called for `Secret`s
(previous commit), we also have to properly handle it in the handler.
Also, handling for `ConfigMap`s is added and the shoot restriction
webhook config augmented such that the handler is actually called when
`ConfigMap`s are created via `gardenadm`.

This is actually missing in the `master` branch already since `gardenadm
connect` already creates these resources here: https://github.com/gardener/gardener/blob/9d121983041a20a1adb57d340d806aa168fc7480/pkg/gardenadm/cmd/connect/connect.go#L224-L236

* Enable `BackupBucket` controller in `gardenlet`

And restrict watch cache to the expected object name.

* Make `.spec.seedName` optional in `BackupBucket`

* Prepare shoot authorizer/restriction for `BackupEntry` controller enablement

* Enable `BackupEntry` controller in `gardenlet`

* Drop superfluent `Get` call for `BackupEntry`

Introduced with https://github.com/gardener/gardener/pull/4179 but not
present in any other component, this doesn't seem to be needed in the
current codebase anymore.

* Augment e2e test

* Reset `.metadata.resourceVersion` before creation

See https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/13519/pull-gardener-e2e-kind-migration/1993203232750440448

* Address PR review feedback

* Introduce `.spec.shootRef` in `Backup{Bucket,Entry}`

* [make generate]

* Use new `.spec.shootRef` field

* Introduce new field selectors for `.spec.shootRef`

and drop `.status.uid` field selector for `Shoot`s (no longer needed
now)

* Adapt authorizer