Add `gardenlet` and `gardener-operator` migrations to perform empty VPA patches on startup (#13573)
* feat(grm/webhook/vpainplaceupdates): add `mutated` label on handler completion

This commit adds a new "mutated" label and instruments the GRM webhook
logic to append it as a final stap of it's hander logic.

The main purpose of this label is to indicate that the resource has
already been processed by the webhook.

This new label will find purpose in a gardenlet initial migration logic
that looks into non-processed VPA resources and applies an empty update,
thus re-triggering the GRM webhook and migrating the pending VPA
resources.

* feat(grm/webhook/vpainplaceupdates/test): verify `mutated` label is added on successful webhook invocation

* feat(cmd/gardenlet): add a new `migration` that performs empty VPA patches on startup

This commit adds a new migration to gardenlet, so that eligible VPA resources,
taht are not yet migrated to InPlaceOrRecreate mode, get picked up by
the GRM webhook.

* feat(cmd/gardener-operator): add a new `migration` that performs empty VPA patches on startup

This commit adds a new migration to the gardener-operator, that applies
an empty Patch update to VPA resources. The main purpose of these
patches, is to allow eligible VPA resources, that need to be
instrumented with updateMode InPlaceOrRecreate, get picked up by the
dedicated GRM webhook.

This migration ensures that Seeds, configured with "alwaysUpdate=false`,
will get their VPA resources picked by the GRM webhook and get their
update modes set appropriately.

* feat(docs/concepts/grm): update `in-place` Pod resource updates webhook section with `mutated` label

* feat(cmd): add cleanup `todo`(s) to `gardenlet` and `gardener-operator` migration

* chore(pkg/grm/webhook): use `metav1.SetMetaDataLabel` to add `mutated` label to `vpainplaceupdates` webhoook

* chore(cmd/internal/migration): move `gardenlet` and `gardener-operator` migration to dedicated `internal` package

This commit moves the empty vpa patches migration logic to a dedicated
"cmd/internal" package so that it's logic can be reused and not duplicated.

Also:
- Use the `features` package to check if `VPAInPlaceUpdates` gate is enabled
- Use the `metav1` package to check if `vpa` resources have correct labels
- Use `metav1` and `v1beta1constants` for `kube-system` and `kubernetes-dashboard` namespaces

* chore(cmd/internal/migration): rename the returned list of VPAs `var` to `vpaList`

* chore(cmd): import reorder by `make generate`

* feat(skaffold): update `skaffold` dependencies with `cmd/internal`

* chore(cmd/internal/migration): move VPA related migrations to `vpa.go`

* feat(cmd/internal/migration): add a migration to apply VPA update Recreate

This commit adds a new migration that applies update mode Recreate to
VPA resources that have been modified by a GRM webhook, activated via
VPAInPlaceUpdates feature gate.

In addition, instrument the gardenlet and gardener-operator components
to run the migration if the VPAInPlaceUpdates feature gate is disabled,
thus undoing the change from the GRM webhook.

* chore(pkg/grm/webhook/in-place): attach `mutated` label to updated VPA resources

* chore(cmd/internal/migration): patch only VPA resources with `Auto` and `Recreate` modes

This commit updates the migration to send patch requests to VPA
resources that are instrumented with update mode Auto and Recreate only.
As this is the behaviour of the GRM webhook, the migration should follow
it's practive and disregard the rest of the updat mode options.

Also:
- Use `client.RawPatch()` for the empty patch send to the VPA resources

* feat(cmd/internal/migration): run `MigrateVPAUpdateModeToRecreate` in parallel

This commit instruments the MigrateVPAUpdateModeToRecreate migration
with functionality to run in parallel based on a VPA chunk size.

* feat(cmd/internal/migration): run `MigrateVPAEmptyPatch` in parallel

This commit instruments the MigrateVPAEmptyPatch migration with
functionality to run in parallel based on a chunk size constant.

* chore(cmd/internal/migration/vpa): use constant number of workers to process the resources

* feat(cmd/internal/migration/vpa): use `labelSelector`(s) to filter `vpa` resources

This commit introduces two label selectors for the vpa migrations,
that are used within .List() calls.

* chore(cmd/internal/migration/vpa): extract common processing logic into a shared utility

* chore(cmd/migration): switch to using `client.Client` to query the API server

This commit switches from using a manager.Manager to client.Client
directly, as the current state of the migration(s) does not require more
than a client instance.

* chore(cmd/internal/migration): rename utility functions

* chore(test/integration/vpainplaceupdates): check if `mutate` label has value

* Update pkg/apis/resources/v1alpha1/types.go

Co-authored-by: Ismail Alidzhikov <9372594+ialidzhikov@users.noreply.github.com>

* Update cmd/internal/migration/vpa.go

Co-authored-by: Ismail Alidzhikov <9372594+ialidzhikov@users.noreply.github.com>

* Update cmd/internal/migration/vpa.go

Co-authored-by: Ismail Alidzhikov <9372594+ialidzhikov@users.noreply.github.com>

* chore(cmd/internal/migration): rename migrations variables to match config naming

* chore(cmd/internal/migration): improve error messages with `patch` and `migrate` phases

* chore(cmd/gardener-operator): switch to `client.Client` arg when running migrations

* Update cmd/internal/migration/vpa.go

Co-authored-by: Tim Usner <tim.usner@sap.com>

* feat(cmd/internal/migration): increase the number of parallel workers to 50

---------

Co-authored-by: Ismail Alidzhikov <9372594+ialidzhikov@users.noreply.github.com>
Co-authored-by: Tim Usner <tim.usner@sap.com>